requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
python-dotenv==1.0.0
httpx==0.25.2
beautifulsoup4==4.12.2
requests==2.31.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.18
sentence-transformers>=2.3.0
huggingface_hub>=0.20.0
numpy>=1.26.0
pandas>=2.1.4
pytest==7.4.3
pytest-asyncio==0.21.1
python-multipart==0.0.6
aiofiles==23.2.1
openai>=1.6.1